Covid 19 analysis
In this project, I intend to utilize Python libraries like Pandas and Matplotlib to download, clean, dissect and picture the Covid open dataset. This dataset is intriguing in light of the fact that it is profoundly organized but then to some degree untidy in manners that genuinely identify with true issues. According to the linear algorithm used for prediction, the algorithm strongly predicts that there is an inclined tendency to increase the death cases along with the cases in the future. We used the Linear regression method.